product update delete functionlity complete

+++ controllers/productControllers.js
@@ -2,26 +2,6 @@ const Products = require("../models/Products");
 
 const cloudinary = require("../utils/cloudinary");
 
-// Controller for handling the creation of a new product
-const createProduct = async (req, res) => {
-  const { title, price, desc, image, category } = req.body;
-  try {
-    const newProduct = new Products({
-      title,
-      price,
-      desc,
-      image,
-      category,
-    });
-    const savedProduct = await newProduct.save();
-
-    res.status(201).json(savedProduct);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-};
-
 const createNewProduct = async (req, res) => {
   try {
     const { title, price, desc, image, category } = req.body;
@@ -80,37 +60,78 @@ const getProductById = async (req, res) => {
 };
 
 // Controller for deleting a single product by ID
+
 const deleteProductById = async (req, res) => {
   try {
     const id = req.params.id;
+
+    const productToDelete = await Products.findById(id);
+
+    if (!productToDelete) {
+      return res.status(404).json({ error: "Product not found" });
+    }
+
+    if (productToDelete.image && productToDelete.image.public_id) {
+      const destroyResponse = await cloudinary.uploader.destroy(
+        productToDelete.image.public_id
+      );
+
+      if (!destroyResponse.result || destroyResponse.result !== "ok") {
+        return res
+          .status(500)
+          .json({ error: "Failed to delete image from Cloudinary" });
+      }
+    }
 
     const deletedProduct = await Products.findByIdAndDelete(id);
 
     if (!deletedProduct) {
       return res.status(404).json({ error: "Product not found" });
     }
+
     res.status(200).json(deletedProduct);
   } catch (error) {
     console.error(error);
     res.status(500).json({ error: "Internal Server Error" });
   }
 };
+
 // Controller for updating a single product by ID
 const updateProductById = async (req, res) => {
   try {
     const id = req.params.id;
 
-    if (!req.body) {
-      return res.status(400).json({ error: "Update data not provided" });
+    if (req.body.image) {
+      if (req.body.image.public_id) {
+        const destroyResponse = await cloudinary.uploader.destroy(
+          req.body.image.public_id
+        );
+
+        if (destroyResponse.result === "ok") {
+          const uploadResponse = await cloudinary.uploader.upload(
+            req.body.image,
+            {
+              upload_preset: "onlineShop",
+            }
+          );
+
+          if (uploadResponse) {
+            req.body.image = {
+              public_id: uploadResponse.public_id,
+              secure_url: uploadResponse.secure_url,
+            };
+          }
+        } else {
+          return res
+            .status(500)
+            .json({ error: "Failed to delete existing image from Cloudinary" });
+        }
+      }
     }
 
     const updatedProduct = await Products.findByIdAndUpdate(id, req.body, {
       new: true,
     });
-
-    if (!updatedProduct) {
-      return res.status(404).json({ error: "Product not found" });
-    }
 
     res.status(200).json(updatedProduct);
   } catch (error) {
@@ -122,7 +143,7 @@ const updateProductById = async (req, res) => {
 module.exports = {
   getProductById,
   deleteProductById,
-  createProduct,
+
   updateProductById,
   getAllProducts,
   createNewProduct,
